music_downloader/requirements.txt
spotipy
yt-dlp
youtube-search
tqdm
python-dotenv
